# frontendmobileapp/swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
  "paths": {
...
    "/api/media/{media_id}": {
      "get": {
        "responses": {
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Media not found"
          },
          "200": {
            "description": "Media file returned"
          }
        },
        "description": "Get media file by ID",
        "operationId": "get_get_media",
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Unique media identifier"
          }
        ],
        "tags": [
          "media"
        ]
      },
      "delete": {
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Media not found"
          },
          "200": {
            "description": "Media deleted successfully"
          }
        },
        "description": "Delete media file",
        "operationId": "delete_get_media",
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Unique media identifier"
          }
        ],
        "tags": [
          "media"
        ]
      }
    },
...
    "/api/profile/about/{keycloak_id}": {
      "parameters": [
        {
          "name": "keycloak_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AboutResponse"
            }
          }
        },
        "summary": "Get user contact & education details for About section",
        "operationId": "get_about",
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/email_notifications": {
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Update email notification settings",
        "operationId": "post_email_notifications",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/EmailNotificationSettings"
            }
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "profile"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/EmailNotificationSettings"
            }
          }
        },
        "summary": "Get email notification settings",
        "operationId": "get_email_notifications",
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/profile/{keycloak_id}": {
      "parameters": [
        {
          "name": "keycloak_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserProfile"
            }
          }
        },
        "summary": "Get user profile",
        "operationId": "get_user_profile",
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/profile_visibility": {
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Update profile visibility settings",
        "operationId": "post_profile_visibility",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ProfileVisibilitySettings"
            }
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "profile"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ProfileVisibilitySettings"
            }
          }
        },
        "summary": "Get profile visibility settings",
        "operationId": "get_profile_visibility",
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/timeline/{keycloak_id}": {
      "parameters": [
        {
          "name": "keycloak_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "responses": {
          "200": {
            "description": "Timeline response containing user post with pagination",
            "schema": {
              "$ref": "#/definitions/TimelineResponse"
            }
          }
        },
        "summary": "retrieve all posts authored by a particular user to  displayed on timeline",
        "operationId": "get_user_timeline",
        "parameters": [
          {
            "type": "integer",
            "default": 20,
            "in": "query",
            "description": "Number of posts per page",
            "name": "per_page"
          },
          {
            "type": "integer",
            "default": 1,
            "in": "query",
            "description": "Page number for pagination",
            "name": "page"
          }
        ],
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/update_profile": {
      "put": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Update user profile",
        "operationId": "put_update_profile",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserProfile"
            }
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/user/keycloak_id": {
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Return authenticated user's Keycloak ID directly from JWT",
        "description": "This endpoint must NOT depend on database state.",
        "operationId": "post_get_user_keycloak_id",
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "profile"
        ]
      }
    },
...
  },
  "info": {
    "title": "YesLove API",
    "version": "1.0"
  },
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
  "definitions": {
    "UserProfile": {
      "properties": {
        "user_id": {
          "type": "integer",
          "description": "User's database ID"
        },
        "username": {
          "type": "string",
          "description": "User's username"
        },
        "bio": {
          "type": "string",
          "description": "User bio"
        },
        "profile_pic": {
          "type": "string",
          "description": "Profile picture URL"
        },
        "user_type": {
          "type": "string",
          "description": "User type: professional or standard"
        },
        "user_posts": {
          "type": "integer",
          "description": "Number of Posts from User"
        },
        "user_followers": {
          "type": "integer",
          "description": "Number of User followers"
        },
        "user_following": {
          "type": "integer",
          "description": "Number of other users followed by current User"
        },
        "contact_info": {
          "$ref": "#/definitions/ContactInfo"
        },
        "education_info": {
          "$ref": "#/definitions/EducationInfo"
        }
      },
      "type": "object"
    },
    "ContactInfo": {
      "properties": {
        "name": {
          "type": "string",
          "description": "User's full name"
        },
        "email": {
          "type": "string",
          "description": "Email address"
        },
        "phone": {
          "type": "string",
          "description": "Phone number"
        },
        "address": {
          "type": "string",
          "description": "User's address"
        },
        "website": {
          "type": "string",
          "description": "User's personal website"
        }
      },
      "type": "object"
    },
    "EducationInfo": {
      "properties": {
        "birthday": {
          "type": "string",
          "description": "User's birthday"
        },
        "education": {
          "type": "string",
          "description": "Education details"
        },
        "institution": {
          "type": "string",
          "description": "Institution name"
        },
        "employment": {
          "type": "string",
          "description": "Current employment details"
        }
      },
      "type": "object"
    },
    "AboutResponse": {
      "properties": {
        "contact": {
          "$ref": "#/definitions/ContactInfo"
        },
        "education_and_employment": {
          "$ref": "#/definitions/EducationInfo"
        }
      },
      "type": "object"
    },
    "ProfileVisibilitySettings": {
      "properties": {
        "settings": {
          "type": "array",
          "description": "List of profile visibility settings",
          "items": {
            "$ref": "#/definitions/ProfileVisibilitySetting"
          }
        }
      },
      "type": "object"
    },
    "ProfileVisibilitySetting": {
      "required": [
        "category",
        "setting_id",
        "value"
      ],
      "properties": {
        "setting_id": {
          "type": "string",
          "description": "Unique ID for the setting"
        },
        "value": {
          "type": "string",
          "description": "Visibility value (visible/hidden)"
        },
        "category": {
          "type": "string",
          "description": "Category: 'Contact' or 'Education And Other Information'"
        }
      },
      "type": "object"
    },
    "EmailNotificationSettings": {
      "properties": {
        "settings": {
          "type": "array",
          "description": "List of email notification settings",
          "items": {
            "$ref": "#/definitions/EmailNotificationSetting"
          }
        }
      },
      "type": "object"
    },
    "EmailNotificationSetting": {
      "required": [
        "setting_id",
        "value"
      ],
      "properties": {
        "setting_id": {
          "type": "string",
          "description": "Unique ID for the setting"
        },
        "value": {
          "type": "boolean",
          "description": "Notification enabled (true/false)"
        }
      },
      "type": "object"
    },
    "TimelineResponse": {
      "properties": {
        "total": {
          "type": "integer",
          "description": "Total number of posts"
        },
        "per_page": {
          "type": "integer",
          "description": "Number of posts per page"
        },
        "current_page": {
          "type": "integer",
          "description": "Current page number"
        },
        "posts": {
          "type": "array",
          "description": "List of posts",
          "items": {
            "$ref": "#/definitions/UserPost"
          }
        }
      },
      "type": "object"
    },
    "UserPost": {
      "properties": {
        "id": {
          "type": "integer",
          "description": "Post ID"
        },
        "author": {
          "type": "string",
          "description": "Post author display name"
        },
        "author_id": {
          "type": "string",
          "description": "Post author keycloak id"
        },
        "author_pic": {
          "type": "string",
          "description": "URL to author's profile picture"
        },
        "content": {
          "type": "string",
          "description": "Text content of the post"
        },
        "image": {
          "type": "string",
          "description": "Legacy image URL field used by detailed post UI"
        },
        "image_url": {
          "type": "string",
          "description": "URL to image in the post"
        },
        "video_url": {
          "type": "string",
          "description": "URL to video in the post"
        },
        "timestamp": {
          "type": "string",
          "description": "Timestamp of the post in ISO format"
        },
        "likes": {
          "type": "integer",
          "description": "Number of likes"
        },
        "comments": {
          "type": "integer",
          "description": "Number of comments"
        },
        "media_files": {
          "type": "array",
          "description": "List of media file URLs associated with the post",
          "items": {
            "$ref": "#/definitions/MediaFile"
          }
        },
        "current_user_reaction": {
          "type": "string",
          "description": "Current user reaction type on the post"
        }
      },
      "type": "object"
    },
    "MediaFile": {
      "properties": {
        "id": {
          "type": "string",
          "description": "Unique media ID"
        },
        "url": {
          "type": "string",
          "description": "Media URL (local or S3)"
        },
        "media_url": {
          "type": "string",
          "description": "Media URL alias for client compatibility"
        },
        "filename": {
          "type": "string",
          "description": "Original filename"
        },
        "content_type": {
          "type": "string",
          "description": "MIME type of the media"
        },
        "file_size": {
          "type": "integer",
          "description": "Size of the file in bytes"
        },
        "width": {
          "type": "integer",
          "description": "Media width in pixels"
        },
        "height": {
          "type": "integer",
          "description": "Media height in pixels"
        },
        "duration": {
          "type": "integer",
          "description": "Media duration in seconds (for video/audio)"
        },
        "created_at": {
          "type": "string",
          "description": "Timestamp when the media was uploaded"
        }
      },
      "type": "object"
    },
...
  },
...
}